exlang/exa/my-study/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "jupyterlab>=4.4.9",
    "langchain>=0.3.27",
    "langchain-ollama",
    "langchain-openai>=0.3.30",
    "langchain-tavily>=0.2.12",
    "python-dotenv>=1.1.1",
]
